chore: 클래스명 rename
#151
